Extract PEP 249 compliant sql_base ext_pillar from MySQL.
Refactor the mysql ext_pillar so all PEP 249 generic code is in
sql_base to allow for easy addition of new ext_pillars.
