+++ tests/unit/pillar/mysql_test.py
@@ -19,7 +19,7 @@ class MysqlPillarTestCase(TestCase):
     maxDiff = None
 
     def test_001_extract_queries_legacy(self):
-        return_data = mysql.Merger()
+        return_data = mysql.MySQLExtPillar()
         args, kwargs = [], {'mysql_query': 'SELECT blah'}
         qbuffer = return_data.extract_queries(args, kwargs)
         self.assertEqual([
@@ -28,7 +28,7 @@ class MysqlPillarTestCase(TestCase):
         ], qbuffer)
 
     def test_002_extract_queries_list(self):
-        return_data = mysql.Merger()
+        return_data = mysql.MySQLExtPillar()
         args, kwargs = [
             'SELECT blah',
             'SELECT blah2',
@@ -63,7 +63,7 @@ class MysqlPillarTestCase(TestCase):
         ], qbuffer)
 
     def test_003_extract_queries_kwarg(self):
-        return_data = mysql.Merger()
+        return_data = mysql.MySQLExtPillar()
         args, kwargs = [], {
             '1': 'SELECT blah',
             '2': 'SELECT blah2',
@@ -92,7 +92,7 @@ class MysqlPillarTestCase(TestCase):
         ], qbuffer)
 
     def test_004_extract_queries_mixed(self):
-        return_data = mysql.Merger()
+        return_data = mysql.MySQLExtPillar()
         args, kwargs = [
             'SELECT blah1',
             ('SELECT blah2', 2),
@@ -123,7 +123,7 @@ class MysqlPillarTestCase(TestCase):
 
     def test_005_extract_queries_bogus_list(self):
         # This test is specifically checking that empty queries are dropped
-        return_data = mysql.Merger()
+        return_data = mysql.MySQLExtPillar()
         args, kwargs = [
             'SELECT blah',
             '',
@@ -161,7 +161,7 @@ class MysqlPillarTestCase(TestCase):
 
     def test_006_extract_queries_bogus_kwargs(self):
         # this test is cut down as most of the path matches test_*_bogus_list
-        return_data = mysql.Merger()
+        return_data = mysql.MySQLExtPillar()
         args, kwargs = [], {
             '1': 'SELECT blah',
             '2': '',
@@ -176,14 +176,14 @@ class MysqlPillarTestCase(TestCase):
         ], qbuffer)
 
     def test_011_enter_root(self):
-        return_data = mysql.Merger()
+        return_data = mysql.MySQLExtPillar()
         return_data.enter_root("test")
         self.assertEqual(return_data.result["test"], return_data.focus)
         return_data.enter_root(None)
         self.assertEqual(return_data.result, return_data.focus)
 
     def test_021_process_fields(self):
-        return_data = mysql.Merger()
+        return_data = mysql.MySQLExtPillar()
         return_data.process_fields(['a', 'b'], 0)
         self.assertEqual(return_data.num_fields, 2)
         self.assertEqual(return_data.depth, 1)
@@ -207,7 +207,7 @@ class MysqlPillarTestCase(TestCase):
         self.assertEqual(return_data.depth, 3)
 
     def test_111_process_results_legacy(self):
-        return_data = mysql.Merger()
+        return_data = mysql.MySQLExtPillar()
         return_data.process_fields(['a', 'b'], 0)
         return_data.with_lists = []
         return_data.process_results([[1, 2]])
@@ -217,7 +217,7 @@ class MysqlPillarTestCase(TestCase):
         )
 
     def test_112_process_results_legacy_multiple(self):
-        return_data = mysql.Merger()
+        return_data = mysql.MySQLExtPillar()
         return_data.process_fields(['a', 'b'], 0)
         return_data.with_lists = []
         return_data.process_results([[1, 2], [3, 4], [5, 6]])
@@ -227,7 +227,7 @@ class MysqlPillarTestCase(TestCase):
         )
 
     def test_121_process_results_depth_0(self):
-        return_data = mysql.Merger()
+        return_data = mysql.MySQLExtPillar()
         return_data.process_fields(['a', 'b', 'c', 'd'], 0)
         return_data.with_lists = []
         return_data.enter_root(None)
@@ -238,7 +238,7 @@ class MysqlPillarTestCase(TestCase):
         )
 
     def test_122_process_results_depth_1(self):
-        return_data = mysql.Merger()
+        return_data = mysql.MySQLExtPillar()
         return_data.process_fields(['a', 'b', 'c', 'd'], 1)
         return_data.with_lists = []
         return_data.enter_root(None)
@@ -249,7 +249,7 @@ class MysqlPillarTestCase(TestCase):
         )
 
     def test_123_process_results_depth_2(self):
-        return_data = mysql.Merger()
+        return_data = mysql.MySQLExtPillar()
         return_data.process_fields(['a', 'b', 'c', 'd'], 2)
         return_data.with_lists = []
         return_data.enter_root(None)
@@ -260,7 +260,7 @@ class MysqlPillarTestCase(TestCase):
         )
 
     def test_124_process_results_depth_3(self):
-        return_data = mysql.Merger()
+        return_data = mysql.MySQLExtPillar()
         return_data.process_fields(['a', 'b', 'c', 'd'], 3)
         return_data.with_lists = []
         return_data.enter_root(None)
@@ -271,7 +271,7 @@ class MysqlPillarTestCase(TestCase):
         )
 
     def test_125_process_results_depth_4(self):
-        return_data = mysql.Merger()
+        return_data = mysql.MySQLExtPillar()
         return_data.process_fields(['a', 'b', 'c', 'd'], 4)
         return_data.with_lists = []
         return_data.enter_root(None)
@@ -282,7 +282,7 @@ class MysqlPillarTestCase(TestCase):
         )
 
     def test_131_process_results_overwrite_legacy_multiple(self):
-        return_data = mysql.Merger()
+        return_data = mysql.MySQLExtPillar()
         return_data.process_fields(['a', 'b'], 0)
         return_data.with_lists = []
         return_data.process_results([[1, 2], [3, 4], [1, 6]])
@@ -292,7 +292,7 @@ class MysqlPillarTestCase(TestCase):
         )
 
     def test_132_process_results_merge_depth_0(self):
-        return_data = mysql.Merger()
+        return_data = mysql.MySQLExtPillar()
         return_data.process_fields(['a', 'b', 'c', 'd'], 0)
         return_data.with_lists = []
         return_data.enter_root(None)
@@ -303,7 +303,7 @@ class MysqlPillarTestCase(TestCase):
         )
 
     def test_133_process_results_overwrite_depth_0(self):
-        return_data = mysql.Merger()
+        return_data = mysql.MySQLExtPillar()
         return_data.process_fields(['a', 'b', 'c', 'd'], 0)
         return_data.with_lists = []
         return_data.enter_root(None)
@@ -314,7 +314,7 @@ class MysqlPillarTestCase(TestCase):
         )
 
     def test_134_process_results_deepmerge_depth_0(self):
-        return_data = mysql.Merger()
+        return_data = mysql.MySQLExtPillar()
         return_data.process_fields(['a', 'b', 'c', 'd'], 0)
         return_data.with_lists = []
         return_data.enter_root(None)
@@ -325,7 +325,7 @@ class MysqlPillarTestCase(TestCase):
         )
 
     def test_135_process_results_overwrite_depth_1(self):
-        return_data = mysql.Merger()
+        return_data = mysql.MySQLExtPillar()
         return_data.process_fields(['a', 'b', 'c', 'd'], 1)
         return_data.with_lists = []
         return_data.enter_root(None)
@@ -336,7 +336,7 @@ class MysqlPillarTestCase(TestCase):
         )
 
     def test_136_process_results_merge_depth_2(self):
-        return_data = mysql.Merger()
+        return_data = mysql.MySQLExtPillar()
         return_data.process_fields(['a', 'b', 'c', 'd'], 2)
         return_data.with_lists = []
         return_data.enter_root(None)
@@ -347,7 +347,7 @@ class MysqlPillarTestCase(TestCase):
         )
 
     def test_137_process_results_overwrite_depth_2(self):
-        return_data = mysql.Merger()
+        return_data = mysql.MySQLExtPillar()
         return_data.process_fields(['a', 'b', 'c', 'd'], 2)
         return_data.with_lists = []
         return_data.enter_root(None)
@@ -358,7 +358,7 @@ class MysqlPillarTestCase(TestCase):
         )
 
     def test_201_process_results_complexity_multiresults(self):
-        return_data = mysql.Merger()
+        return_data = mysql.MySQLExtPillar()
         return_data.process_fields(['a', 'b', 'c', 'd'], 2)
         return_data.with_lists = []
         return_data.enter_root(None)
@@ -370,7 +370,7 @@ class MysqlPillarTestCase(TestCase):
         )
 
     def test_202_process_results_complexity_as_list(self):
-        return_data = mysql.Merger()
+        return_data = mysql.MySQLExtPillar()
         return_data.process_fields(['a', 'b', 'c', 'd'], 2)
         return_data.with_lists = []
         return_data.enter_root(None)
@@ -383,7 +383,7 @@ class MysqlPillarTestCase(TestCase):
         )
 
     def test_203_process_results_complexity_as_list_deeper(self):
-        return_data = mysql.Merger()
+        return_data = mysql.MySQLExtPillar()
         return_data.process_fields(['a', 'b', 'c', 'd'], 0)
         return_data.with_lists = []
         return_data.enter_root(None)
@@ -396,7 +396,7 @@ class MysqlPillarTestCase(TestCase):
         )
 
     def test_204_process_results_complexity_as_list_mismatch_depth(self):
-        return_data = mysql.Merger()
+        return_data = mysql.MySQLExtPillar()
         return_data.as_list = True
         return_data.with_lists = []
         return_data.enter_root(None)
@@ -411,7 +411,7 @@ class MysqlPillarTestCase(TestCase):
         )
 
     def test_205_process_results_complexity_as_list_mismatch_depth_reversed(self):
-        return_data = mysql.Merger()
+        return_data = mysql.MySQLExtPillar()
         return_data.as_list = True
         return_data.with_lists = []
         return_data.enter_root(None)
@@ -427,7 +427,7 @@ class MysqlPillarTestCase(TestCase):
         )
 
     def test_206_process_results_complexity_as_list_mismatch_depth_weird_order(self):
-        return_data = mysql.Merger()
+        return_data = mysql.MySQLExtPillar()
         return_data.as_list = True
         return_data.with_lists = []
         return_data.enter_root(None)
@@ -445,7 +445,7 @@ class MysqlPillarTestCase(TestCase):
         )
 
     def test_207_process_results_complexity_collision_mismatch_depth(self):
-        return_data = mysql.Merger()
+        return_data = mysql.MySQLExtPillar()
         return_data.as_list = False
         return_data.with_lists = []
         return_data.enter_root(None)
@@ -460,7 +460,7 @@ class MysqlPillarTestCase(TestCase):
         )
 
     def test_208_process_results_complexity_collision_mismatch_depth_reversed(self):
-        return_data = mysql.Merger()
+        return_data = mysql.MySQLExtPillar()
         return_data.as_list = False
         return_data.with_lists = []
         return_data.enter_root(None)
@@ -476,7 +476,7 @@ class MysqlPillarTestCase(TestCase):
         )
 
     def test_209_process_results_complexity_collision_mismatch_depth_weird_order(self):
-        return_data = mysql.Merger()
+        return_data = mysql.MySQLExtPillar()
         return_data.as_list = False
         return_data.with_lists = []
         return_data.enter_root(None)
@@ -494,7 +494,7 @@ class MysqlPillarTestCase(TestCase):
         )
 
     def test_20A_process_results_complexity_as_list_vary(self):
-        return_data = mysql.Merger()
+        return_data = mysql.MySQLExtPillar()
         return_data.as_list = True
         return_data.with_lists = []
         return_data.enter_root(None)
@@ -511,7 +511,7 @@ class MysqlPillarTestCase(TestCase):
         )
 
     def test_207_process_results_complexity_roots_collision(self):
-        return_data = mysql.Merger()
+        return_data = mysql.MySQLExtPillar()
         return_data.as_list = False
         return_data.with_lists = []
         return_data.enter_root(None)
@@ -525,7 +525,7 @@ class MysqlPillarTestCase(TestCase):
         )
 
     def test_301_process_results_with_lists(self):
-        return_data = mysql.Merger()
+        return_data = mysql.MySQLExtPillar()
         return_data.as_list = False
         return_data.with_lists = [1, 3]
         return_data.enter_root(None)
@@ -550,7 +550,7 @@ class MysqlPillarTestCase(TestCase):
         )
 
     def test_302_process_results_with_lists_consecutive(self):
-        return_data = mysql.Merger()
+        return_data = mysql.MySQLExtPillar()
         return_data.as_list = False
         return_data.with_lists = [1, 2, 3]
         return_data.enter_root(None)
